+++ shared/types/api/socket-events.ts
@@ -1,0 +1,74 @@
+import type { Server as SuperServer, Socket as SuperServerSocket } from 'socket.io';
+import type { Socket as SuperSocket } from 'socket.io-client';
+
+import type { TRoomState } from '../room';
+import type { TRoomsList } from '../room/list';
+import type { VisualGameState } from '../game/state';
+import type { GameOptions } from '../game/options';
+import type { TVoteOption } from '../game/vote';
+import type { TMissionResult } from '../game/mission';
+import type { TRoles } from '../game/roles';
+import type { TAssassinateType } from '../game/addons';
+import type { IRoomUnavailableError } from './errors';
+import type { TTotalWinrateStats } from '../stats';
+import type { PublicUserProfile } from '../user';
+
+export type {
+  ISocketError,
+  IRoomUnavailableError,
+  ILoginError,
+  IRegisterError,
+  IUpdateEmailError,
+  IUpdatePasswordError,
+  IUpdateLoginError,
+  IUpdateAvatarError,
+} from './errors';
+
+export type { ArgumentOfCallback } from './helpers';
+
+export interface ServerToClientEvents {
+  roomsListUpdated: (list: TRoomsList) => void;
+  onlineCounterUpdated: (counter: number) => void;
+  roomOnlineUpdated: (counter: number) => void;
+  roomUpdated: (state: TRoomState) => void;
+  gameUpdated: (state: VisualGameState) => void;
+  restartGame: (uuid: string) => void;
+  destroyRoom: (uuid: string) => void;
+  serverError: (error: string) => void;
+  renewJWT: () => void;
+}
+
+export interface ClientToServerEvents {
+  getTotalStats: (callback: (stats: TTotalWinrateStats) => void) => void;
+  getPlayerGames: (uuid: string, callback: (games: VisualGameState[]) => void) => void;
+  getRoomsList: (callback: (list: TRoomsList) => void) => void;
+  getOnlineCounter: (id: string, callback: (counter: number) => void) => void;
+  getUserProfile: (id: string, callback: (user: PublicUserProfile) => void) => void;
+
+  createRoom: (callback: (uuid: string) => void) => void;
+  updateOptions: (uuid: string, options: GameOptions) => void;
+  joinRoom: (uuid: string, callback: (state: TRoomState | IRoomUnavailableError) => void) => void;
+  lockRoom: (uuid: string) => void;
+  kickPlayer: (uuid: string, userID: string) => void;
+  leaveRoom: (uuid: string) => void;
+
+  endGame: (uuid: string) => void;
+  endAndRestartGame: (uuid: string) => void;
+  shuffle: (uuid: string) => void;
+  voteInRoom: (uuid: string, result: boolean) => void;
+
+  joinGame: (uuid: string) => void;
+  startGame: (uuid: string) => void;
+  leaveGame: (uuid: string) => void;
+  restartGame: (uuid: string) => void;
+
+  selectPlayer: (uuid: string, userID: string) => void;
+  sentSelectedPlayers: (uuid: string) => void;
+  voteForMission: (uuid: string, option: TVoteOption) => void;
+  actionOnMission: (uuid: string, result: TMissionResult) => void;
+  assassinate: (uuid: string, type: TAssassinateType, role?: TRoles) => void;
+}
+
+export type Server = SuperServer<ClientToServerEvents, ServerToClientEvents>;
+export type Socket = SuperSocket<ServerToClientEvents, ClientToServerEvents>;
+export type ServerSocket = SuperServerSocket<ClientToServerEvents, ServerToClientEvents>;