+++ shared/types/game/mission.ts
@@ -1,0 +1,12 @@
+export type TMissionResult = 'success' | 'fail';
+
+export interface MissionSettings {
+  players: number;
+  failsRequired: number;
+}
+
+export interface MissionWithResult extends MissionSettings {
+  hidden?: boolean;
+  result?: TMissionResult;
+  fails?: number;
+}