+++ shared/types/game/vote.ts
@@ -1,0 +1,11 @@
+export type TVoteOption = 'approve' | 'reject';
+
+export interface TeamMember {
+  id: string;
+  preVote?: TVoteOption;
+}
+
+export interface PreVoteData {
+  playerID: string;
+  value: TVoteOption;
+}