+++ shared/types/stats/index.ts
@@ -1,0 +1,19 @@
+import type { TOptionalRoles } from '../game/roles';
+
+export type TTotalWinrateStats = {
+  total: TWinrateStats;
+  byPlayers: (TWinrateStats & { playerCount: number })[];
+  roleStats: TRoleStats[];
+};
+
+export type TWinrateStats = {
+  gamesCount: number;
+  goodWins: number;
+  evilWins: number;
+  goodWinPercentage: number;
+  evilWinPercentage: number;
+};
+
+export type TRoleStats = {
+  role: TOptionalRoles;
+} & TWinrateStats;