+++ shared/types/user/avatars.ts
@@ -1,0 +1,5 @@
+export interface IAvatarInfo {
+  id: string;
+  available: boolean;
+  info?: string;
+}